requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.30.6
pytest==8.3.2
pytest-django==4.8.0
pytest-asyncio==0.23.8
redis==5.0.8
django-prometheus==2.3.1
django-redis==5.4.0

